moved fetch-by-id fixtures
